BROKEN -- first pass of resource inheritance, just to look at on phone
