+++ js/Resource.js
@@ -1,9 +1,5 @@
 // @ts-check
 
-import { deepCopy, fix } from "./Utils.js";
-import { settings } from "./Settings.js";
-import { resourceConfigs } from "./configs/ResourceConfigs.js";
-
 class Resource {
     constructor ({
         internalName,
@@ -11,10 +7,9 @@ class Resource {
         displayNamePlural,
         flavorText,
         amount,
-        hitpoints,
         active,
         resourceDiv,
-        world,
+        container,
     }) {
         this.internalName = internalName;
         this.displayNameSingular = displayNameSingular;
@@ -23,11 +18,9 @@ class Resource {
         // TODO: amount vs. quantity
         this.amount = amount;
         this.amountPerTick = 0;
-        this.hitpoints = hitpoints;
-        this.isFocused = false;
         this.active = active;
         this.resourceDiv = resourceDiv;
-        this.world = world;
+        this.container = container;
 
         if (this.active) {
             this.constructDOM();
@@ -40,7 +33,6 @@ class Resource {
             return;
         }
         this.amountDiv = document.createElement("div");
-        this.amountDiv.onclick = () => {this.world.resources.setFocusedResource(this)};
         this.amountDiv.classList.add("tooltip");
         this.amountSpan = document.createElement("span");
         this.amountDiv.appendChild(this.amountSpan);
@@ -53,70 +45,15 @@ class Resource {
         this.resourceDiv.appendChild(this.amountDiv);
     }
 
-    draw () {
-        if (!this.active) {
-            if (this.amount > 0) {
-                this.active = true;
-                this.constructDOM();
-            }
-            else {
-                return;
-            }
-        }
-
-        const fixedAmount = fix(this.amount);
-        const nameToUse = fixedAmount === 1 ? this.displayNameSingular : this.displayNamePlural;
-        const amountPerSecond = fix(this.amountPerTick * settings.fps * 100) / 100;
-
-        const plus = amountPerSecond >= 0 ? "+" : ""; // + if amountPerSecond positive, nothing if negative (it'll have its own negative)
-        const newTooltipSpanHTML = `<span class="tooltipTextInner">${this.flavorText}<br><br>Currently: ${plus}${amountPerSecond} per second<hr></span>`;
-        if (this.tooltipSpan.innerHTML !== newTooltipSpanHTML) {
-            this.tooltipSpan.innerHTML = newTooltipSpanHTML;
-        }
-
-        this.amountSpan.textContent = `${fixedAmount} ${nameToUse} (${plus}${amountPerSecond}/sec)`;
-
-        // UI change if current resource is focused
-        if (this.isFocused) {
-            this.amountSpan.style.fontWeight = 'bold';
-        }
-        else {
-            this.amountSpan.style.fontWeight = '';
-        }
-    }
-
-    // To keep track of the resource gain per tick (and consequently per second),
-    // use startTick to zero out the gain, and then for adding or consuming the
-    // resource, use tickAdd and tickConsume.  When buying (or other actions
-    // that shouldn't be tracked in resource gain per second, like selling),
-    // use noTickAdd and noTickConsume.
-    startTick () {
-        this.amountPerTick = 0;
-    }
-
-    tickAdd (amount) {
-        this.amount += amount;
-        this.amountPerTick += amount;
-    }
-
-    tickConsume (amount) {
-        this.amount -= amount;
-        this.amountPerTick -= amount;
-    }
-
-    noTickAdd (amount) {
+    add (amount) {
         this.amount += amount;
     }
 
-    noTickConsume (amount) {
+    consume (amount) {
         this.amount -= amount;
     }
 
-    tickFocus (amount) {
-        this.tickAdd(amount / this.hitpoints / settings.fps);
-    }
-
-    // Saving (loading is at the bottom with createResource)
+    // Saving (loading is special because it depends on the config)
     save () {
         const save = {};
         save.n = this.internalName;
@@ -127,19 +64,4 @@ class Resource {
     }
 }
 
-
-function createResource (name, resourceDiv, world) {
-    return new Resource({ ...deepCopy(resourceConfigs[name]), resourceDiv: resourceDiv, world: world });
-}
-
-function loadResource (save, resourceDiv, world) {
-    const config = deepCopy(resourceConfigs[save.n]);
-    config.amount = save.am;
-    config.active = save.ac === 1 ? true : false;
-    config.resourceDiv = resourceDiv;
-    config.world = world;
-
-    return new Resource(config);
-}
-
-export { createResource, loadResource };
+export { Resource };
